requirements.txt
magika~=0.5.1
PyYAML~=6.0.1